BG-2347: Merge partition data

+++ yupana-core/src/main/scala/org/yupana/core/model/MetricData.scala
@@ -16,4 +16,4 @@
 
 package org.yupana.core.model
 
-case class MetricData(count: Long, time: Double, speed: Double)
+case class MetricData(count: Long, time: Long, speed: Double)
